Created Tester classes for each API and added tests to it
